pyproject.toml
[project]
name = "quantumalpha"
version = "0.1.0"
description = "AI 原生量化投研决策系统 - 机构级精度的跨平台桌面端智能投研工作台"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "QuantumAlpha Team" }]

dependencies = [
    # 数据源
    "tushare>=1.4.0",
    "akshare>=1.14.0",
    
    # 数据处理
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    
    # 技术分析（注意：TA-Lib 需要系统级安装）
    "ta>=0.11.0",  # 纯 Python 技术分析库，作为 TA-Lib 的替代
    
    # HTTP 客户端
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    
    # 数据验证
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # LLM 集成
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",
    
    # 通知服务
    "aiosmtplib>=3.0.0",
    
    # 日志与配置
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    
    # 异步任务调度
    "apscheduler>=3.10.0",
    
    # 向量存储 (RAG) - 暂时禁用，Python 3.14 兼容性问题
    # "chromadb>=0.5.0",
    
    # 前后端通信
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
